feat: :construction: routes->controller->service->database pattern
