+++ config/test.ts
@@ -1,0 +1,19 @@
+export default {
+	port: 4000,
+	dbUrl: '',
+	jwtSecret: 'daDSBSJDBJSBD324542NKDNKSNDKND',
+	jwtExpiry: '3d',
+	cloudinaryName: '',
+	cloudinaryApiKey: '',
+	cloudinaryApiSecret: '',
+	smtpHost: '',
+	smtpPort: '',
+	smtpUser: '',
+	smtpPass: '',
+	stripeApiKey: '',
+	stripeApiSecret: '',
+	razorpayApiKey: '',
+	razorpayApiSecret: '',
+	userImageDir: 'users',
+	productImageDir: 'products'
+};
