+++ jest.config.js
@@ -1,0 +1,11 @@
+/** @type {import('ts-jest/dist/types').InitialOptionsTsJest} */
+module.exports = {
+	preset: 'ts-jest',
+	testEnvironment: 'node',
+	testMatch: ['**/**/*.test.ts'],
+	verbose: true,
+	forceExit: true,
+	clearMocks: true,
+	resetMocks: true,
+	restoreMocks: true
+};
