+++ package.json
@@ -9,11 +9,11 @@
 	},
 	"scripts": {
 		"dev": "nodemon -r tsconfig-paths/register ./src/index.ts",
+		"test": "jest",
 		"start": "node -r tsconfig-paths/register ./build/src/index.js",
 		"build": "yarn clean && tsc",
 		"postbuild": "yarn run copy-files",
 		"clean": "rm -rf build/*.js build/*.map build/*",
-		"copy-files": "copyfiles src/swagger.yaml build/",
 		"format": "prettier --write \"./**/*.ts\"",
 		"prepare": "husky install",
 		"lint": "eslint src",
@@ -38,11 +38,13 @@
 		"@types/express": "^4.17.13",
 		"@types/express-fileupload": "^1.2.0",
 		"@types/express-pino-logger": "^4.0.3",
+		"@types/jest": "^29.0.0",
 		"@types/jsonwebtoken": "^8.5.6",
 		"@types/nanoid": "^3.0.0",
 		"@types/node": "^16.11.12",
 		"@types/nodemailer": "^6.4.4",
 		"@types/stripe": "^8.0.417",
+		"@types/supertest": "^2.0.12",
 		"@types/swagger-ui-express": "^4.1.3",
 		"@types/yamljs": "^0.2.31",
 		"@typescript-eslint/eslint-plugin": "^5.8.0",
@@ -55,10 +57,13 @@
 		"eslint-plugin-import": "^2.25.3",
 		"eslint-plugin-prettier": "^4.0.0",
 		"eslint-plugin-security": "^1.4.0",
-		"husky": "^7.0.0",
+		"husky": "^8.0.0",
+		"jest": "^29.0.1",
 		"lint-staged": "^12.1.4",
 		"nodemon": "^2.0.15",
 		"prettier": "^2.5.1",
+		"supertest": "^6.2.4",
+		"ts-jest": "^28.0.8",
 		"ts-node-dev": "^2.0.0",
 		"typescript": "^4.7.4"
 	},
