--- src/schema/user/index.ts
@@ -1,7 +1,0 @@
-/* eslint-disable import/prefer-default-export */
-export {default as registerUserSchema} from './registerUser.schema';
-export {default as loginUserSchema} from './loginUser.schema';
-export {default as forgotPasswordSchema} from './forgotPassword.schema';
-export {default as passwordResetSchema} from './resetPassword.schema';
-export {default as changePasswordSchema} from './changePassword.schema';
-export {default as updateUserSchema} from './updateUser.schema';
